# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Interview Management API
  description: REST API for managing interviews, authentication, and resources.
  version: 1.0.0
servers:
  - url: http://localhost:5001/api
    description: Local development server
...
paths:
...
  /interviews:
    post:
      tags: [Interviews]
      summary: Create a new interview
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Interview'
      responses:
        '200':
          description: Interview created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  interview: { $ref: '#/components/schemas/Interview' }

    get:
      tags: [Interviews]
      summary: Get all interviews (with optional status filter)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, completed]
      responses:
        '200':
          description: List of interviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  interviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/Interview'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Interview:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        date: { type: string, format: date }
        time: { type: string }
        interviewer: { type: string }
        duration: { type: string }
        notes: { type: string }
        status: { type: string, enum: [upcoming, completed] }
...